+++ models/builderModels/models-other/Lesson.js
@@ -6,40 +6,22 @@ const config = require('../../../real_config.js')
 const Series = require('./Series.js')
 
 var lessonSchema = new Schema({
-    bibleRefs: [config.bibleRefSchema],
-    collaborators: {
-        type: [String],
-        default: []
-    },
-    content: {
-        application: {
-            type: String,
-            default: ''
-        },
-        hook: {
-            type: String,
-            default: ''
-        },
-        main: {
-            type: String,
-            default: ''
-        },
-        prayer: {
-            type: String,
-            default: ''
-        }
+    bibleRefs: [config.bibleRefSchema], // refs that the lesson is primarily focused on, not all refs used in the lesson
+    createdBy: {
+        type: String,
+        required: true
     },
     mainIdea: {
         type: String,
         default: ''
     },
-    media: [config.mediaRefSchema],
+    media: [config.mediaRefSchema], // any media added from the builder side, not associated with a research object
     research: [config.researchRefSchema],
     seriesID: {
         type: Schema.Types.ObjectId,
         ref: 'series'
     },
-    tags: {
+    tags: { // used in searching for lessons later
         type: [String],
         default: []
     },
@@ -47,9 +29,9 @@ var lessonSchema = new Schema({
         type: String,
         required: true
     },
-    user: {
-        type: String,
-        required: true
+    users: {
+        type: [String],
+        default: []
     }
 }, { toJSON: { virtuals: true } })